refactor(control-panels): split useControlPanelsLogic into tab hooks
Extract backupShared, appearance/account/system/sync hooks, and thin
composer. Update error-boundary wiring test to read useSystemSettings.
